--- src/main/java/com/landvibe/landlog/controller/MemberForm.java
@@ -1,13 +1,0 @@
-package com.landvibe.landlog.controller;
-
-public class MemberForm {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
