+++ src/main/java/com/landvibe/landlog/controller/MemberLoginForm.java
@@ -1,0 +1,22 @@
+package com.landvibe.landlog.controller;
+
+public class MemberLoginForm {
+    private String email;
+    private String password;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
