+++ src/main/java/com/landvibe/landlog/controller/MemberNewForm.java
@@ -1,0 +1,31 @@
+package com.landvibe.landlog.controller;
+
+public class MemberNewForm {
+    private String name;
+    private String email;
+    private String password;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
